+++ libs/oef-core/src/comms/public_key_utils.cpp
@@ -48,7 +48,7 @@ inline std::string RSAKey::to_string() const
     return out;
   }
   char *        pk_ptr = nullptr;
-  unsigned long len    = BIO_get_mem_data(mem, &pk_ptr);
+  unsigned long len    = static_cast<unsigned long>(BIO_get_mem_data(mem, &pk_ptr));
   if (!len)
   {
     FETCH_LOG_WARN(LOGGING_NAME_PK, " while getting bio char pointer");
@@ -87,7 +87,7 @@ inline std::string RSAKey::to_string_base64() const
     return out;
   }
   char *        pk_ptr = nullptr;
-  unsigned long len    = BIO_get_mem_data(mem, &pk_ptr);
+  unsigned long len    = static_cast<unsigned long>(BIO_get_mem_data(mem, &pk_ptr));
   if (!len)
   {
     FETCH_LOG_WARN(LOGGING_NAME_PK, " while getting bio char pointer");
